# heatmap/docs/api_spec.yaml
openapi: "3.1.0"
info:
  title: Jordan Kindergarten Heatmap API
  version: "1.0.0"
  description: |
    REST API for the Jordan National Kindergarten Intelligence Heat Map.
    Provides GeoJSON layers, composite indicators, statistical analytics,
    ETL pipeline control, and multi-channel alerts.
servers:
  - url: /api/heatmap
    description: Mounted at FastAPI root
...
paths:
  /geojson:
    get:
      tags: [geojson]
      summary: Jordan admin GeoJSON with indicator overlays
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
          description: Filter by specific date (YYYY-MM-DD); defaults to latest
      responses:
        "200":
          description: GeoJSON FeatureCollection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoJSONCollection"
...
  /indicators:
    get:
      tags: [indicators]
      summary: Composite indicator values
      parameters:
        - name: admin_id
          in: query
          schema: { type: string }
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: List of indicator records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IndicatorRecord"
...
  /alerts:
    get:
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: unacknowledged_only
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{alert_id}/acknowledge:
    post:
      tags: [alerts]
      summary: Acknowledge an alert
      parameters:
        - name: alert_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
...
components:
  schemas:
...
    IndicatorRecord:
      type: object
      properties:
        date:                  { type: string, format: date }
        admin_id:              { type: string }
        kindergarten_status:   { type: number }
        children_enrollment:   { type: number }
        staff_classrooms:      { type: number }
        safety_incidents:      { type: number }
        reports_attendance:    { type: number }
        tasks_governance:      { type: number }
...
    Alert:
      type: object
      properties:
        id:           { type: string, format: uuid }
        severity:     { type: string, enum: [HIGH, MEDIUM, INFO] }
        admin_id:     { type: string }
        rule:         { type: string }
        message:      { type: string }
        meta:         { type: object }
        created_at:   { type: string, format: date-time }
        acknowledged: { type: boolean }
...
    GeoJSONCollection:
      type: object
      properties:
        type:     { type: string, enum: [FeatureCollection] }
        features: { type: array, items: { type: object } }